requirements.txt
# bu ikisini yukleyin

python-dotenv==0.19.0

social-auth-app-django==5.0.0

